Frontend/requirements.txt
imblearn==0.0
requests==2.31.0
dvc==2.58.2
mlflow==2.7.1
deepchecks
pytest
streamlit==1.25.0
matplotlib==3.8.0
numpy==1.25.2
pandas==1.3.5
pillow==9.5.0

